Move common exception classes to their own files
