+++ projectq/backends/__init__.py
@@ -31,6 +31,7 @@ This includes:
 from ._aqt import AQTBackend
 from ._awsbraket import AWSBraketBackend
 from ._circuits import CircuitDrawer, CircuitDrawerMatplotlib
+from ._exceptions import DeviceNotHandledError, DeviceOfflineError, DeviceTooSmall
 from ._ibm import IBMBackend
 from ._ionq import IonQBackend
 from ._printer import CommandPrinter
