+++ projectq/meta/_exceptions.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+#   Copyright 2021 ProjectQ-Framework (www.projectq.ch)
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+"""Define exception classes for projectq.meta."""
+
+
+class QubitManagementError(Exception):
+    """
+    Exception raised when the lifetime of a qubit is problematic within a context manager.
+
+    This may occur within Loop, Dagger or Compute regions.
+    """
