+++ projectq/meta/_loop.py
@@ -28,11 +28,8 @@ from copy import deepcopy
 from projectq.cengines import BasicEngine
 from projectq.ops import Allocate, Deallocate
 
+from ._exceptions import QubitManagementError
 from ._util import drop_engine_after, insert_engine
-
-
-class QubitManagementError(Exception):
-    """Exception raised when the lifetime of a qubit is problematic within a loop."""
 
 
 class LoopTag:  # pylint: disable=too-few-public-methods
